reset resets count on each route
