+++ src/client.js
@@ -1,11 +1,11 @@
 'use strict';
 
-const FetchMock = require('./fetch-mock');
+const fetchMock = require('./fetch-mock');
 const statusTextMap = require('./status-text');
 const theGlobal = typeof window !== 'undefined' ? window : self;
 
-module.exports = new FetchMock({
-	theGlobal: theGlobal,
+module.exports = fetchMock({
+	global: theGlobal,
 	Request: theGlobal.Request,
 	Response: theGlobal.Response,
 	Headers: theGlobal.Headers,
