+++ src/server.js
@@ -4,11 +4,11 @@ const Request = fetch.Request;
 const Response = fetch.Response;
 const Headers = fetch.Headers;
 const stream = require('stream');
-const FetchMock = require('./fetch-mock');
+const fetchMock = require('./fetch-mock');
 const http = require('http');
 
-module.exports = new FetchMock({
-	theGlobal: global,
+module.exports = fetchMock({
+	global: global,
 	Request: Request,
 	Response: Response,
 	Headers: Headers,
